+++ src/app/auth/models/user-service.model.ts
@@ -1,0 +1,5 @@
+import {SettingsModel} from '../../core/models/settings.model';
+
+export interface UserServiceModel extends SettingsModel {
+  password: string;
+}