+++ src/app/core/core.module.ts
@@ -4,6 +4,7 @@ import {ConstantService} from './services/constant.service';
 import {Generated, GenerateFactory, GeneratorService} from './services/generator';
 import {ConfigOptionsService} from './services/config-options.service';
 import {SharedModule} from '../shared/shared.module';
+import {LocalStorageService} from './services/local-storage.service';
 
 const constants = new ConstantService({ App: 'Shop', Ver: '1.0' });
 
@@ -32,6 +33,10 @@ const config = new ConfigOptionsService({
       useFactory: GenerateFactory(2),
       deps: [ GeneratorService ]
     },
+    {
+      provide: localStorage,
+      useClass: LocalStorageService
+    }
   ]
 })
 export class CoreModule { }