+++ src/app/core/guards/can-deactivate.guard.ts
@@ -1,0 +1,21 @@
+import { Injectable } from '@angular/core';
+import {CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree, CanDeactivate} from '@angular/router';
+import { Observable } from 'rxjs';
+import {CanComponentDeactivate} from '../interfaces/can-component-deactivate';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class CanDeactivateGuard
+  implements CanDeactivate<CanComponentDeactivate> {
+  canDeactivate(
+    component: CanComponentDeactivate
+  ):
+    | Observable<boolean | UrlTree>
+    | Promise<boolean | UrlTree>
+    | boolean
+    | UrlTree {
+
+    return component.canDeactivate();
+  }
+}