+++ src/app/core/models/settings.model.ts
@@ -1,0 +1,8 @@
+import { UserRole } from './role.enum';
+
+export interface SettingsModel {
+  id: number;
+  login: string;
+  email: string;
+  roles: UserRole[];
+}