+++ src/app/core/services/config-options.service.ts
@@ -1,12 +1,13 @@
 import { Injectable } from '@angular/core';
 import {CoreModule} from '../core.module';
+import {SettingsModel} from '../models/settings.model';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ConfigOptionsService {
-
   private readonly configOptions: object;
+  private settingsStorage: SettingsModel;
 
   constructor(configOptions = {}) {
     this.configOptions = configOptions;
@@ -22,5 +23,13 @@ export class ConfigOptionsService {
 
   getAll(): object {
     return this.configOptions;
+  }
+
+  set settings(newSettings: Partial<SettingsModel>) {
+    this.settingsStorage = { ...this.settingsStorage, ...newSettings};
+  }
+
+  get settings(): Partial<SettingsModel> {
+    return this.settingsStorage;
   }
 }