+++ src/app/products/products.module.ts
@@ -1,8 +1,9 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import {ProductComponent} from './components/product/product.component';
 import {ProductListComponent} from './components/product-list/product-list.component';
 import {SharedModule} from '../shared/shared.module';
+import { ProductDetailsComponent } from './components/product-details/product-details.component';
+import {RouterModule} from '@angular/router';
 
 
 
@@ -10,12 +11,13 @@ import {SharedModule} from '../shared/shared.module';
   declarations: [
     ProductComponent,
     ProductListComponent,
+    ProductDetailsComponent,
   ],
   exports: [
     ProductListComponent,
   ],
   imports: [
-    SharedModule
+    SharedModule, RouterModule
   ]
 })
 export class ProductsModule { }