+++ src/app/products/services/products.service.ts
@@ -8,6 +8,7 @@ import {map} from 'rxjs/operators';
   providedIn: 'root'
 })
 export class ProductsService {
+  private products: IProductModel[] = [ ...products ];
 
   constructor() { }
 
@@ -18,5 +19,25 @@ export class ProductsService {
           return items.map(product => new ProductModel(product));
         })
       );
+  }
+
+  getProduct$(id: number): Observable<ProductModel> {
+    return of(this.products)
+      .pipe(
+        map((items: IProductModel[]) => items.find(product => product.id === id)),
+        map(product => new ProductModel(product))
+      );
+  }
+
+  addProduct(product: IProductModel): void {
+    product.id = +new Date();
+
+    this.products.push(product);
+  }
+
+  editProduct(product: IProductModel): void {
+    const inx = this.products.findIndex(i => i.id === product.id);
+
+    this.products[inx] = { ...this.products[inx], ...product };
   }
 }