+++ src/app/shared/components/select-order/select-order.component.ts
@@ -1,5 +1,10 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 
+export interface IOptionsSelectOrder {
+  value: string;
+  label?: string;
+}
+
 @Component({
   selector: 'app-select-order',
   templateUrl: './select-order.component.html',
@@ -11,7 +16,7 @@ export class SelectOrderComponent implements OnInit {
   initial: string;
 
   @Input()
-  options: { value: string, label?: string }[];
+  options: IOptionsSelectOrder[];
 
   @Input()
   label: string;